Downgraded PHP to match prod env.

+++ database/migrations/2024_02_23_052826_create_stickers_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
     Schema::create('stickers', function (Blueprint $table)
     {
       $table->id();
-      $table->unsignedInteger('recipient_id');
-      $table->unsignedInteger('sender_id');
+      $table->unsignedBigInteger('recipient_id');
+      $table->unsignedBigInteger('sender_id');
       $table->string('status')->default('pending');
       $table->foreign('recipient_id')->references('id')->on('singers')->onDelete('cascade')->onUpdate('cascade');
       $table->foreign('sender_id')->references('id')->on('singers')->onDelete('cascade')->onUpdate('cascade');
